refactor(cost): double-precision spend and round only the top stack segment

+++ src/session/migrations.ts
@@ -125,7 +125,7 @@ export const migrations: Migration[] = [
         model TEXT NOT NULL DEFAULT '',
         prompt_tokens BIGINT NOT NULL DEFAULT 0,
         completion_tokens BIGINT NOT NULL DEFAULT 0,
-        step_cost REAL NOT NULL DEFAULT 0,
+        step_cost DOUBLE PRECISION NOT NULL DEFAULT 0,
         PRIMARY KEY (tenant_id, day, agent, model)
       );
     `,
